+++ packages/ui/src/components/page/Sidebar/CurrentMember/index.ts
@@ -1,0 +1,1 @@
+export * from './CurrentMember'
