+++ packages/ui/src/components/typography/index.ts
@@ -1,3 +1,4 @@
+export * from './Badge'
 export * from './Label'
 export * from './Text'
 export * from './TokenValue'
